requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pycryptodome
paho-mqtt
pyjwt
requests
cryptography